LayoutTests/fast/page-color-sampling/color-sampling-includes-subframes-expected.txt

PASS Loaded iframe
PASS color.top is "rgb(255, 100, 0)"
PASS successfullyParsed is true

TEST COMPLETE
